app/requirements.txt
aioredis==1.3.1
redis==3.5.3
discord.py==1.6.0
python-telegram-bot==13.1